Fix a merge issue with dry run, and move stats to their own file in pygcp.bigquery. Fixes a merge issue with setup.py.
